workers/pyproject.toml
[project]
name = "photodisplay-workers"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "boto3==1.34.162",
  "exifread==3.0.0",
  "httpx==0.27.0",
  "pillow==10.4.0"
]
